Update the supported library set for Flutter for web (#39983)
