+++ packages/flutter_tools/lib/src/build_runner/build_script.dart
@@ -41,6 +41,21 @@ const String jsSourceMapExtension = '.ddc.js.map';
 const String kReleaseFlag = 'release';
 const String kProfileFlag = 'profile';
 
+// A minimum set of libraries to skip checks for to keep the examples compiling
+// until we make a decision on whether to support dart:io on the web.
+// See https://github.com/dart-lang/sdk/issues/35969
+// See https://github.com/flutter/flutter/issues/39998
+const Set<String> skipPlatformCheckPackages = <String>{
+  'flutter',
+  'flutter_test',
+  'flutter_driver',
+  'flutter_goldens',
+  'flutter_goldens_client',
+  'flutter_gallery',
+  'connectivity',
+  'video_player',
+};
+
 final DartPlatform flutterWebPlatform =
     DartPlatform.register('flutter_web', <String>[
   'async',
@@ -127,7 +142,7 @@ final List<core.BuilderApplication> builders = <core.BuilderApplication>[
               sdkKernelPath: path.join('kernel', 'flutter_ddc_sdk.dill'),
               outputExtension: ddcKernelExtension,
               platform: flutterWebPlatform,
-              librariesPath: 'libraries.json',
+              librariesPath: path.absolute(path.join(builderOptions.config['flutterWebSdk'], 'libraries.json')),
               kernelTargetName: 'ddc',
             ),
         (BuilderOptions builderOptions) => DevCompilerBuilder(
@@ -135,7 +150,7 @@ final List<core.BuilderApplication> builders = <core.BuilderApplication>[
               platform: flutterWebPlatform,
               platformSdk: builderOptions.config['flutterWebSdk'],
               sdkKernelPath: path.url.join('kernel', 'flutter_ddc_sdk.dill'),
-              librariesPath: 'libraries.json',
+              librariesPath: path.absolute(path.join(builderOptions.config['flutterWebSdk'], 'libraries.json')),
             ),
       ],
       core.toAllPackages(),
@@ -201,7 +216,8 @@ class FlutterWebTestEntrypointBuilder implements Builder {
   @override
   Future<void> build(BuildStep buildStep) async {
     log.info('building for target ${buildStep.inputId.path}');
-    await bootstrapDdc(buildStep, platform: flutterWebPlatform);
+    await bootstrapDdc(buildStep, platform: flutterWebPlatform,
+        skipPlatformCheckPackages: skipPlatformCheckPackages);
   }
 }
 
@@ -229,7 +245,8 @@ class FlutterWebEntrypointBuilder implements Builder {
     if (release || profile) {
       await bootstrapDart2Js(buildStep, flutterWebSdk, profile);
     } else {
-      await bootstrapDdc(buildStep, platform: flutterWebPlatform);
+      await bootstrapDdc(buildStep, platform: flutterWebPlatform,
+          skipPlatformCheckPackages: skipPlatformCheckPackages);
     }
   }
 }
