+++ packages/flutter_tools/lib/src/build_runner/web_fs.dart
@@ -422,7 +422,11 @@ class BuildDaemonCreator {
             }
             break;
           default:
-            printTrace(serverLog.message);
+            if (serverLog.message.contains('Skipping compiling')) {
+              printError(serverLog.message);
+            } else {
+              printTrace(serverLog.message);
+            }
         }
       },
       buildMode: daemon.BuildMode.Manual,
